Improve Hive SerDe perf by caching column projection information
Imporoves the Hive SerDe performance by marking caching projection information for
struct fields. We now check if a column is "selected" (part of the projection) and
what the index in the projection is. Previously, we had to resolve the column index
for each row. Note that when the ObjectInspector is initialized we only have the
table schema and don't know the projection schema until the first record is processed.

This provides ~2x speedup over the previous approach for our test queries.

Change-Id: I69c55e6d49cc412dc07aa556f78925a4050e7fee
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6005
Tested-by: jenkins
Reviewed-by: Lenni Kuff <lskuff@cloudera.com>
Tested-by: Lenni Kuff <lskuff@cloudera.com>