Shade dependent jars for client and spark build artifacts.
The spark README shows the result. We only need the user to add a
single jar to the classpath and this has minimal/no code bloat as
the client shaded jar is required anyway.

I think shading thrift is required for other reasons beyond ease of
use. Other components (e.g. spark) use it and the version mismatch
is no fun.

Change-Id: Icb7ef7d34ee80636a6ab65b53d029bf96424f363
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6095
Tested-by: jenkins
Reviewed-by: Nong Li <nong@cloudera.com>