Update benchmarking scripts to allow generating and comparing results not in the DB.
This patch updates the benchmarking and reporting scripts to work without needing
the results to be persisted in the DB. This serves one main use case of benchmarking
a branch that is not committed. We don't want to collect the main results with these
one-off tests.

With these changes, it should be easy to setup a jenkins job that runs a private branch,
pulls in the most recent trunk timings, appends the results from the private run and
then generates the same plots as in the dash board.

Change-Id: Ifb68f4553c38dbde740de8cf1562c8be440e0a7c
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6111
Tested-by: jenkins
Reviewed-by: Lenni Kuff <lskuff@cloudera.com>