Add support for timestamp type.
This is a bit more complex because the parquet and Impala in memory encoding is
really not suitable for the java side. Impala stores it as:
  8-byte time of day in nanos, 4-byte julian day.

The java library support likes:
  8-byte time since epoch in milliseconds, 4 byte nanos.

The java format is also where the less precise types (Datetime, Time, etc) are
going across the platform. This is a marginal perf loss by having Impala do
the conversion but more than makes up for it on the java side.

Change-Id: I0e47bbadaf61cf4543773fe7b4bda1c4b7b5d2ea
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6159
Tested-by: jenkins
Reviewed-by: Nong Li <nong@cloudera.com>