Reduce serialized task size.
On the 500GB cluster tests, the task size is multiple GBs and blows up trying to
serialize the thrift result. This fails since java arrays have a 2GB size limit
(int32 length restriction).

This is indicative of a bigger problem that the tasks are too big. Even if we did
not hit the int limit, this is no good.

This patch does a few things to help, resulting in a 3 orders of magnitude reduction in
total task size on the 500GB table.

1. Avoid splitting the logic between FE and BE. The FE now ships the normal Impala
plan and the BE does the work of generating tasks.
2. Switch to compact thrift serialization.
3. Compress the tasks using lz4.
4. Minimize the per-task plan. This removes the unreferenced hosts and partitions.

We still have an issue that each task contains the entire plan and associated metadata.
We can probably at most bring this down 10x but that is still unlikely to meet the
scaling requirements. We need to think about this some more.

Change-Id: Ifc7f0af6f1d3707f38b2057f4e9254d50ba7b038