Update builds scripts to deal with the split repos.
To build, we first need to build the client repo, then the Impala repo. This
is problematic because the Impala config overwrites THRIFT_HOME to point to
a thirdparty folder inside the Impala repo. Thrift is not built until Impala
is. The client repo needs thrift as well but any stock thrift will do.

This patch splits the build to use system thrift to build the client and then
sets up the Impala env for Impala.

Change-Id: I6e54cbe0b0965c278fc8efdcedacfbc5003a73b1