Cap number of scanner threads when running Impala on RS.
Impala, by default, configures to 3x number of cores so number of threads.
This is much higher than the number of slots. On our cluster, this is higher
than the number of thrift server worker threads (64) causing tasks to get
rejected.

3x number of cores would be very high for number of slots. For now, we'll just
cap it but will need to think about this some more.

Change-Id: I32cdd2642db281cf5d96568cd5a4dc922b12db95
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6413
Tested-by: jenkins
Reviewed-by: Nong Li <nong@cloudera.com>