Updates to service API to better support fault tolerance.
This adds the idea of an ordered result and offset. Ordered means each time
the task is executed, the records returned are in the same order. This with
offset means if an node fails halfway, the client can just remember how many
rows it has seen and continue from that offset. Scans from HDFS are the common
case and they are ordered.

Not all scans are ordered though, for example kudu.

This patch also adds updates to the Client library to automatically retry.
There's more to be done here wrt trying different hosts.

Change-Id: I10be041c40634128502bdab07e9949fdacd937f0
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6640
Reviewed-by: Nong Li <nong@cloudera.com>
Tested-by: Nong Li <nong@cloudera.com>