Add external mini cluster utility.
This adds a utility that creates an external mini cluster (forks the daemon
processes). This seems like a good way for us to do fault testing. We should
extend this so that this can be called from Java.

subprocess.h/.cc is a port from kudu.

This includes one test case that demonstrates how this could be used.

Change-Id: I738a990ce3338303e26947c553d88b4044069cc2
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6698
Tested-by: jenkins
Reviewed-by: Nong Li <nong@cloudera.com>