Add kerberos sanity check test.
This is a very hacked up test case to verify that kerberos is not broken. We probably
will throw all of this away. I've manually set up a jenkins box (vd0232) to be in the
same realm as our kerberos cluster and will set up a jenkins job that just runs on that
machine. This will *not* be part of the normal check in tests.

I've added this as a stop gap to verify the behavior for our alpha customer. There's some
issues with kerberos for the new services that we need to work out.

Change-Id: Id2952d57ffb50ff4dadf4f064ca6c13f9e212881
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6983
Tested-by: jenkins
Reviewed-by: Alexander Leblang <alex.leblang@cloudera.com>