Add thread pool style execution to RS testing
This patch creates a load testing class for the RecordService
in the form of a JobQueue class. The class instantiates a
ThreadPoolExecutor and includes methods that can add tasks to the
executor and check the the Futures returned by the executor have
completed.

Future work includes writing more test cases and adding more
advanced result verification.

This patch also includes a small improvement to TestMiniClusterController

Change-Id: Ia878675b9fde0f0705997f471994f427d864f70b
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7147
Tested-by: jenkins
Reviewed-by: Alexander Leblang <alex.leblang@cloudera.com>