Some tests for encrypting tasks
This adds a few tests for encrypting tasks.

First it tests scenarios where request is planned on a secure cluster
and executed on a non-secured cluster, and the case where request is
planned on a non-secure cluster and executed on a secure cluster.

It also adds a fuzz test where it swaps a random byte with the last byte
in the task binary, and expects the task to fail the HMAC authentication
after a few tries.

Change-Id: I5195357a2eb6c6b4c691c849b5ad61a42e8e3807
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7422
Tested-by: jenkins
Reviewed-by: Chao Sun <chao.sun@cloudera.com>