Make GetProtocolVersion throws TRecordServiceException
This adds exception catching logic to Thrift generated code,
so that on the server side we can check connection limit
inside the GetProtocolVersion method, and throw exception if
the threshold has been reached. The exception will be sent back
to the client side, so the user could get a proper response
for this case.

Change-Id: I2dde8db2a2b6b2b9934f6eaf1c7f354dba134839
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7835
Tested-by: jenkins
Reviewed-by: Chao Sun <chao.sun@cloudera.com>