Update client to build on OSX.
This patch lets up build the thrift API and java client on OSX. More work is necessary
to build the cpp client but we don't expect that to be used very often.

We need to document how to install thrift on OSX but we need to try from a fresh OSX machine.

Change-Id: I2a111b9570644bbedc288b952e8572e3debf7d99
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7867
Tested-by: jenkins
Reviewed-by: Chao Sun <chao.sun@cloudera.com>