Schedule Tasks on Remote DN
Currently, we always pick a worker host where there's a local replica.
This doesn't work very well if the data set is skewed, in which case
a small potion of the hosts will get assigned most of the tasks.

With this, it first tries to schedule the task locally if there's a replica
on the current host. Otherwise, it picks a random worker host from the
global membership. Hopefully this could mitigate the skew better.

Change-Id: Iedb5ea84299f4379b1044241c51566da7d2481d5
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7963
Tested-by: jenkins
Reviewed-by: Chao Sun <chao.sun@cloudera.com>