Jenkins Fixes
This patches contains a few client side changes that fix some
problems with out jenkins builds and allow us to correctly use
amazon burst build machines.

The CLEAN flag was previously broken; when CLEAN was true the
build would run git clean after installing the toolchain which
would then delete the toolchain and cause the bulid to fail.
This patch fixes that, and also always turns the clean flag on
if the build is on an ec2 burst machine. This activates the
-format_metastore flag in the buildall.sh script on the server
side which creates the necessary metastore db.

This patch also uses the impala-python virtualenv when runing
RS server tests which allows us to always have the correct
python packages.

Change-Id: I2422a93b74f24683d19da7e782daa3265bb1a187
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/11655
Reviewed-by: Alexander Leblang <alex.leblang@cloudera.com>
Tested-by: Alexander Leblang <alex.leblang@cloudera.com>