Adding MiniClusterTests back into Project
When we reorganized the file structure for the beta we got rid of the
minicluster tests. This patch adds in a new directory for the tests
and fixes a longstanding bugs in the minicluster. Because the minicluster
cannot run on a node that already contains a recordserviced running, the
pom file has been changed to exclude the minicluster tests from the main
mvn test run.
Future work includes creating a target to run the minicluster tests so
that they will also be run on precommit.

Change-Id: I2aab7049ae6a0fa43dc8e7a571329a58b38473e1
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/9018
Tested-by: jenkins
Reviewed-by: Chao Sun <chao.sun@cloudera.com>