RS-29: Support planner auto discovery (part 1)
This patch adds the option of ZK planner discovery for MR jobs.
This mode can be enabled by specifying "recordservice.zookeeper.connectString"
in the job configuration. Once enabled, when generating splits, the MR job
will query ZK service and get a list of planner addresses to use.

Note that, the ZK test added here cannot be run when the test user has kerberos
credentials.

TODO: add the support for Spark jobs as well.

Change-Id: I87fb9435823c89173bd08b5de953c7cf5133a8b4
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/10682
Tested-by: Jenkins User
Reviewed-by: Chao Sun <chao.sun@cloudera.com>
Tested-by: Chao Sun <chao.sun@cloudera.com>
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/11874